Fix calculation of number of micropolygon sub-bounds for motion sampling.
The number of sub-bounds should be roughly equal to the total number of
micropolygon lengths which are moved through during the time interval.
The previous calculation was flawed in two ways:
1) It divided by the shading rate, which is the *area* of the micropolygon, not
   the *length*.
2) It assumed that the shading rate was actually achieved.  This can be far
   from true for large shading rates and small primitives come into play, since
   the shading rate is essentially capped by the primitive size.

The new calculation uses the size of the bound at the first key frame to
estimate the size of the micropolygon rather than the desired shading rate, and
should be more robust.  It results in significant speedups for motion blurred
scenes with large shading rates, which would previously have resulted in too
few subbounds.

The shadingRate member was removed from SqGridInfo because it is no longer
needed after these changes.