Motionfactor Optimization
ri.cpp
=======
Replaced some stub code which maps the motionfactor value to some structure sitting in memory.

attributes.cpp
===============
Added an ADD_ATTR.. for GeometricMotionFactor, setting default val = 1.0

bound.h/.cpp
============
Defined and implemented getBoundCuboid which fills the passed array with the points of a bounding box.

surface.h/.cpp
===============
Added a big block of code to AdjustedShadingRate().  This code is only activated if objects are moving (more on that later).  It uses the approach of calculating the minimum speed, throughout all of the transformations in the motion-block, of the object in question and then using that value * the total exposure time * (a magic constant) to scale the ShadingRate.  The idea is that if an object is moving rather fast it will be heavily blurred and we can get away with cranking up the ShadingRate because small surface details won't be visible anyway.

Also to surface.h, added an isMoving() which calls isMoving on its own transformation and returns the result.

transform.h
============
Added a getter, isMoving() returns the private m_IsMoving