Fix motionfactor motion estimation code to work correctly.
The previous motion estimation was unfortunately buggy due to the use of
m_Bound, which is in an odd hybrid camera/raster coordinate system, with
transformation bound added.  The use of axis-aligned bounds rather than
cubeoids for estimating the translation is also inaccurate and was replaced.

The code works and provides a speed increase but several things are quite
suboptimal, such as the recomputation of the object bound.