Bugfix for bound adjustment with transformation motion.
In order to adjust the bound to encapsulate motion transformations, the bound
in object space needs to be computed from the camera-space bound.  The object
is stored in camera space at time = 0, so the transformation back to object
space should also happen at time = 0 rather than at the first key frame time of
the object transformation.