Bugfix 2827244: Large numbers of RiPoints incorrectly bound
The bounding code was broken for large numbers of RiPoints: whenever a points
collection was split into two collections, the member m_MaxWidth was not
correctly initialized to the maximum width of the points in camera space.  As a
result the bounding was incorrect, with large points being placed into the
wrong bucket and incorrectly sampled.

Thanks to Ricardo Mayor for the initial work on this bug.