Bugfix 2811280: Incorrect incident ray length in orthogonal projection
The builtin shader variable "I" (incident ray direction) should have a length
which is equal to the distance between the point P and the ray origin.  The ray
origin in the orthographic case is somewhere on the xy plane, so "I" should be
set to (0,0,P.z()).

This behaviour is consistent with other renderman renderers as discussed in an
aside in the usenet thread "E should be varying"
(comp.graphics.rendering.renderman, 2002)

Initial investigation and patch courtesy of Ricardo Mayor.