Improved storage mechanism for array primvars.
The storage mechanism for data inside a CqParameterTypedVaryingArray used to be
std::vector<std::vector<T>>, but this is likely to be quite inefficient for
small array lengths, since it implies an excessive number of dynamic
allocations and bad memory locality.

The alternative implemented here is to store the array of arrays as a flat
std::vector<T>, and do the 2D indexing calculation by hand.  It avoids the
dynamic allocation problems and allows the pValue() function to return a
pointer to all the data at once.