Minor optimizations + bugfix for CqPoints::NaturalDice to allow array primvars.
CqPoints::NaturalDice had a bug which caused array primvars to be incorrectly
copied from the CqParameter into the shader data array, which is now fixed.

The optimizations to CqPoints::NaturalDice and CqPoints::Bound make sure we
work with the raw parameter and shader data arrays, to avoid an excessive
number of virtual function calls.  In a scene with lots of points the speedup
is modest however; perhaps only 1% or so.