Bugfix 2812953: shadow("blur") filter artifacts with large values
This was due to the way we approximate surfaces during percentage closer
filtering using a linear approximation.  While a linear approximation is
superior for very smooth surfaces, it is unpredictable for bumpy surfaces,
irrelevant for volume shading and cannot be used for faked subsurface
scattering.

This patch replaces the linear approximation with a constant depth
approximation for predictability and conformance with other renderers.  A new
filtering parameter for the shadow() shadop, "depthapprox" is introduced which
may be set to either "constant" (the default) or "linear" to turn on the linear
approximation for smooth surfaces.