Bugfix 2846751: Camera motion blur not working with points
This bug was caused by omission of the camera key frames when determining the
key frames needed for transformation motion blur of points primitives.  The
patch rectifies this, introducing a function mergeKeyTimes to merge the key
frame times from two transformation objects into a vector of times.  Also
included is some minor cleanup of CqMicroPolyGridPoints::Split() related to the
use of a vector of times.

A peripheral bug regarding at which time matCameraToObject0 was extracted was
also fixed.