Bugfix 2853241: min() not working properly with varying
This bug was due to using the function argument cParams as a loop counter
inside the varying loop, which only worked correctly for the first varying
element.  The bug was present in all of SO_min, SO_max, SO_pmin, SO_pmax,
SO_cmin, SO_cmax; the patch rectifies the problem in all locations.