Bugfix 2853236: No AOV for volume shaders
Arbitrary output variables were not copied from volume shaders onto output
grids, and therefore not sampled during the sampling stage.

This fix makes sure that AOVs from volume shaders are accessible by looking for
the AOV name on the atmosphere shader whenever it's not found in the surface
shader associated with a surface.