Convert all files ending with CRLF to LF endings.
git has the ability to check out files with platform-specific line endings
(CRLF pairs on windows, LF on unix) based on a configuation setting.  For this
to work properly all files stored in the repository need to be stored with
plain LF endings; this change ensures that is the case.