Bugfix 2133952: smooth shading interpolation artefacts
The shading artifacts were due to an incompatibility between smooth shading and
one-sided derivatives for normal computations.  The problem arises because
one-sided derivatives are asymmetric about a vertex, but for smooth shading we
need to compute normals accurately at the *vertices*.  On the other hand, for                                                                                                    constant interpolation, we want to compute normals on the *faces* of the
micropolygons which is naturally achieved using one-sided differences.

This patch introduces centred differences as required to fix the bug, but
leaves one-sided differences in place for use with constant shading
interpolation.  The type of derivatives to use is automatically deduced from
the type of shading interpolation.  A new attribute, "derivatives" "centered"
is introduced to allow manual control of the type of derivatives in use.

Details:
* Introduce a new class for computing derivatives on regular grids, CqGridDiff.
  (This migntn't be the best way to go about computing derivatives, but it will
  do for now.  Perhaps the grids should instead "know" how to compute
  derivatives themselves.)

* Remove the old derivative computations SO_DuType and SO_DvType, which were
  only used by CqMicroPolyGrid::CalcSurfaceDerivatives().

* Reimplement
  - CqMicroPolyGrid::CalcNormals(),
  - CqMicroPolyGrid::CalcSurfaceDerivatives() and
  - CqShaderExecEnv::{diffU,diffV}
  to make use of CqGridDiff.  This means that all derivatives should now be
  consistent between the shaderVM and the core.

* Make CqShaderExecEnv initialize the CqGridDiff with the "derivatives"
  "centred" attribute to control use of centred differences.

* Add "uniform int centered" to the default token dictionary.