Use a relative tolerance for detecting when N comes out too small.
When numerically computing the normal N based on surface tangents, we need to
detect when N has come out "near zero" due to grid degeneracies, so that we can
use some neighbouring grid points in an attempt to get a more numerically
robust value for N.

Using an absolute tolerance isn't really good enough for this task, since in
that case the decision about whether N is "near zero" depends on the absolute
size of the geometry.  This patch makes sure we use a relative tolerance,
chosen based on the length of the grid diagonal.