Bugfix 2838827: crop command causes artifacts
Filtering artifacts appeared at the edges of a cropped region, due to a bad
interaction between the bucket overlap caching and "rendering" of buckets which
were fully outside of the crop region:  The sample region of a bucket could be
reduced in side due to application of a cache from a non-rendered bucket, and
then reduced in size again due to the crop window.  This meant that some                samples required for filtering were not present at the top and left edges of            the crop window.

This change makes sure that we never "render" buckets which lie fully outside
the crop window at all.

Details:
* Introduce a CqRegion defining the valid buckets, CqImageBuffer::m_bucketRegion
* Make sure PostSurface, RepostSurface, AddMPG and NextBucket are adjusted to             use the new bucket region rather than the full non-cropped set of buckets.
* Disable the general bucket order code in NextBucket for now since it's not
  being used, and definitely not ready for prime time even if it was enabled.
* The order of SetImage() and initialiseCropWindow() in ri.cpp needed to be
  changed.