Make micropolygon edge equations more numerically robust
It's very important that a sample point near the boundary between two
micropolys lies in exactly one or the other micropolygon.  If it hits neither
then a hole appears; if it hits both then a darker spot will be apparent.

The edge equations therefore have to be constructed in exactly the same way for
any given micropolygon and its neighbour; this includes using the same "point
on the line" as well as the same edge vector.  This change satisfies the
requirement by using the midpoint of the edge as the point on the line.