Allow rendering of multiple grid types
It would be nice to render multiple different grid types with the same
piece of code, but to have all the grid-specific stuff inlined into the
rasterization function.  This patch implements one half of the visitor
pattern (grids are now polymorphic) to do just that.  The other half
will come if we end up with multiple Renderer implementations.

- Mostly just code reorginisation to cut down on interdependencies between
  headers.
- Renamed Grid -> QuadGrid to make way for other grid types.
- Fix a bunch of utility functions which should have been declared
  inline but weren't.