Add directory for experimental code & edge eqn test
Understanding the reasons for choosing a particular version of the edge
equations (for point-in-polygon testing) is tricky.  edge_eqn_test.cpp contains
some tests which show why the fastest version isn't the best.

Also created a general "experiments" directory to contain small one-off
experimental programs & move the sse experiments to it as well.