Changes the point in poly edge tests for numerical robustness.
As it turns out, using the edge midpoint for the point-in-poly edge tests isn't
any good: it fails to guarentee that the end points of a given edge actually
lie on the edge as determined by the edge equation!  To fix this we now use one
of the edge end points for the required point on the line (this is the method
used in aqsis).  There's possibly some other problems with this method, but it
seems to work quite well in practise.