Refactor sampling loop for per-grid initialization
This refactoring gives us the opportunity to have a per-grid sampling
initialization stage for the rasterization.

It also fixes some bogus gcc compiler warnings regarding uninitialized
data.  Fixing these in the straightforward way (by simply initializing
everything to zero) seems to have a minor but measureable performance
hit.

To make the reference "simple" sampler more self-contained, the sampler
object abstraction was removed from it and everything was hardcoded into
the loop over the micropolys in a grid.  (As a side effect, this means
it doesn't have sufficient abstraction to deal with point micropolys,
but that's no problem for a special purpose reference implementation.)