AOV mechanism for sample storage and image output
This change allows arbitrary AOVs to be selected via the Renderer
interface, and makes sure that they're correctly passed along to the
image output.  A separate tiff image is saved for each AOV, with the AOV
name appended to the file.

Also create a set of standard VarSpecs and move remaining nontrivial
methods from renderer.h into renderer.cpp