Fix bugs related to uniform variables & arrays
This change fixes some bugs related to handling of uniform data and
arrays.  FvecView needed to get a new member, m_elSize, to record the
length of the contained vector elements, which is different from the
stride in the uniform case.  The storage size for array data was also
being incorrectly computed, so this is fixed as well.

Finally, added stream insertion for VarSpec's for debugging convenience.