Optimization: Precompute hashes for shader variable names
Patch from Brent Fulgham:

One thing that consistently shows up as a high cost area is the computation of
hash values for the variable values, specifically the 'FindLocalVarIndex'
methods in shadervm.h.  Precomputing the hash values seems to be a big win
here, particularly for shaders with large numbers of local variables.