Refactor grid storage variable handling.
This change ensures that we can have different variables attached to a
grid compared to the set of parent primvars.  This is important, because
the renderer might want to fill in extra variables from default values
present in the attribute state (eg, Cs).  Alternatively, the some
geometric variables may be special (eg, P on a sphere) and therefore not
present in the primvar set of the geometry.

Details:
  * Refactor GridvarStorage, renaming to GridStorage.  It's now
    constructed by creating an instance of GridStorageBuilder and adding
    gridvars as necessary before calling the build() method.
  * Add boolean testing to FvecView & DataView for checking when they're
    holding a valid pointer.  Also add a uniform() method to FvecView to
    check whether it represents uniform storage.
  * Fix some errors in stdvar declarations & declare s & t.
  * Move StdvarIndices into varspec.h
  * Implement VarSet, a class representing a sorted set of variables
    with fast lookup via binary_search, and keeping track of the indices
    of standard variables.
  * Add some tests for GridStorage, VarSpec and VarSet