Rearrange scenes for easier benchmarking
In an attempt to make various revisions easier to compare, the scene
definitions have been separated from main.cpp and put into their own
files.  Tenpatch is a simple scene which stresses the rasterizer with
lots of patches, while the default scene isn't a benchmark, it's more of
a demo.