Refactor StdvarIndices struct to enable iteration.
The way the StdvarIndices struct was set up, it was impossible to
iterate over all the standard indices.  This change sets things up so
that adding iteration will be easier.

Also, make sure -Wall is actually used to compile, and make storage() a
virtual grid method for near future convenience.