Improve variable flow; particularly for std vars
This patch improves the handling of primitive and grid variables, to
make it somewhat more correct and standard conforming:
- Ensure that vector and normal type primvars are transformed correctly
- Ensure that N and Ng are correctly diced.

Also update the default scene to use interpolated normals for
displacement to avoid cracks and to use a prettier displacement
function.  The cube mesh is now added to the renderer via a somewhat
generic quad mesh interpreter which generates a bunch of patches.