Implement an iterator for visiting image samples
SampleStorage::Iterator abstracts the process of visiting samples away
into an iterator, which keeps the details of SampleStorage more
self-contained.  As a bonus, this also seems to give some performance
improvements because the iterator may be optimized to avoid some
arithmetic while maintaining the SampleStorage encapsulation.

Also switched to using util.h clamp() and floor() implementations for
notational convenience & control.