Consolidate array view stuff into a single primary template
The FvecView class, the DataView class and their associated const versions were
substantially the same.  This change refactors them so that the common parts
are consolidated into a templated base mixin class which can be used via the
curiously recurring template pattern.  The remaining functionality is mostly
the various forms of constructors which are implemented in the derived classes.