Attempt at bugfix in windows shader compilation
The expected path to the shader compiler was hardcoded in the changes
made for RSLCompile, which was fine on the linux build, but broke things
on windows.  This changes things so that the _target_ name aqsl is
specified in the custom command, which causes cmake to automatically
determine the executable location.

It also modifies the semantics of RSL_COMPILER for users of RSLCompile;
it's now reserved for fully qualified paths.  RSL_COMPILER_NAME can be
used instead to invoke automated lookup with find_program().