Refactor GeomHolder with deformation in mind.
This change makes GeomHolder into a class, with the idea that it will
gain slightly more complicated behaviour as deformation motion blur is
implemented.

I'm using intrusive_ptr here because if shared_ptr is used, we end up
with a few percent performance hit (tested at ~4% percent on core2) with
the tenpatch test scene.