Optimization: intrusive_ptr rather than shared_ptr
shared_ptr is needlessly inefficient for small objects which will be
created in large numbers, because (1) it requires a separate heap
allocated reference count and (2) the required storage is larger than a
pointer.

This change replaces all use of shared_ptr with intrusive_ptr and
introduces a non-virtual counted base, "RefCounted" to ease the creation
of objects with reference counts.  The performance improvements for the
tenpatch scene are almost 10%.