Sample generation experiments.
This experimental code attempts to generate samples which are
well-distributed in both time and space, for use with interleaved
sampling.  It uses Lagae's corner tiles coupled with a new (?) method
for generating reasonable sample distributions inside each tile using
the avaliable interleaving times.  Low noise seems to be achievable,
even with interleaving tiles which are a non integer multiple of the
pixel size.