Implement DoF for sample generation prototype
This change implements optimization of sample patterns including both
time and lens offsets.  DoF sampling of a disc is compared to the exact
result to get the standard deviation for quantitative purposes.

Some artifacts appear, probably due to the limited tile set size of 16
tiles.  I need to investigate increasing the number of corner colours to
3 to conteract this.