Improved sample stratification with (MB xor DoF)
When one of motion blur or depth of field are turned on, but not the
other, we can improve the sample stratification algorithm by ignoring
the sample dimension which is not activated.  This implements that
change, with a noticeable improvement in the variance for scenes with
one of MB or DoF as a result.