Implement atomic reference counting
This change uses implementation details from boost.interprocess to
implement an atomic refcount for use with boost::intrusive_ptr.  Relying
on the implementation details is nasty, but boost will likely have a
nice atomic<T> template which we can replace this with in the future.
Other possibilities involve either bringing in an extra library
dependency, or writing lots of platform-dependent assembler ourselves :(

The change also moves the reference counting machinary into a new file.