Avoid RIB input buffer char translation
This small optimization replaces the integer input buffer with a
narrower one based on unsigned char, avoiding an extra translation
stage.  To do this we need to avoid trying to store EOF (== -1) in the
buffer, which is outside the range of unsigned char.  Instead, we can
consistently use 255 as a proxy for the true EOF, since this is already
used by RIB ProcRunPrograms to indicate the end of stream in any case.
Speedups seem to be on the order of 10% of the pure lexing time.