Flesh out Ri::Filter base, add frame drop filter
This change makes a new version of the Ri::Filter base class - pretty
simple for now, just including basic storage of the renderer services &
next filter in line so we don't have to implement that boilerplate for
each new filter.  Some renaming is also included, which implies changes
to the validation filter.

In addition, the change implements a simple filter, OnOffFilter, which
lets interface calls through if a flag is true; this in turn is used as
a base class to implement a frame dropping filter.  Some details of the
frame dropping don't quite work yet - in particular the parser gets
confused about calls which modify the renderer state like LightSource
and ObjectBegin; needs fixing next.