Create a script to run all code generation
This script executes the codegen for all files in the aqsis source tree
which contain the cog source file tag, "[[[cog".  It sets up the correct
path pointing to ri.xml, and to the utilities file, codegenutils.py.
For convenience it can be executed from any directory inside the source
tree.