RI -> Ri::Renderer refactoring, ri.cpp functions
Modify all Ri* functions from the main ri.cpp file so that they are
methods of a new Ri::Renderer instance called RiCxxCore.  Refactor
various things so that the formal parameters of these methods can be
matched with the implementations, which were written for RI-style
function parameters.

At the moment the code has the odd feature that there are two separate
filter chains, with the C API in the middle.  Obviously this needs to be
fixed in the next stage of refactoring, but for now it's a nice way to
test the C API in ri2ricxx.{h,cpp} which we've got as the new renderer
frontend.