A filter for retained Object/Archive instancing
The filter sits in front of the main renderer API and deals with both
ObjectInstance and InlineArchive API caching.  This commit also removes
the remaining parts of the old object instancing mechanism, and fixes
one or two const-correctness problems with the new cache classes.

Also fix some bugs in the caching code, and improve the docs.