Fix scope validation in validate filter
The scope stack in a validation filter needs to be popped *after*
passing a ScopeEnd command onto the next filter in the layer.  If it's
popped first, delayed procedurals which are processed during the
renderer WorldEnd call will seem to occur outside of the world block,
which would be invalid.