Remove handles from C++ API
The way in which the traditional RI passes handles back to the client
from procedures like RiObjectBegin and LightSource considerably
complicates things like inline archive caching.  Therefore I've removed
handles entirely from the C++ version of the API, making it even more
RIB-like.  This means that the RI->RiCxx translation code needs to
maintain a name->handle mapping internally, but removes the need for
this from most other places in the code.

Details:
* Light sources now held by CqRenderer rather than in a global list
* The way that the Ri->RiCxx context is handled is made to match the C
  API
* Fix validation so that archive scope turns off scope checking
* Add some extra info to ri.xml to support the difference in return
  types between the C and C++ APIs
* Change ri2ricxx.cpp to account for the fact that it now needs to
  generate handles and turn them into names for the C++ API.
* Remove lots of code from RIB parser - it no longer needs to mess with
  name -> handle mappings.
* Improve some parse error message formatting
* Fix various additional problems with inline archive handling