Embed libaqsis_riutil tests in the library
Doing this means that the relevant symbols don't have to be exported on
windows in order for the test executables to correctly compile, which
should simplify things a lot.

The aqsis_add_library cmake macro has been extended with an additional
optional argument, TEST_SOURCES.  If TEST_SOURCES is non-empty and
testing is enabled, then the test sources are compiled into the library,
and a test runner executable is created and added to the set of tests.