Reinstate the -echoapi functionality
With these changes, the -echoapi option will once again write RIB to
stdout, but now with all the functionality and improved robustness of
RibWriter.

To implement this feature, a "tee" filter is introduced, which directs
its input to a _pair_ of output streams.  Also implemented is a very
simple "ignorearchives" filter for dropping ReadArchive requests from
the output (if not, the commands contained within would appear twice).