Change BSD lincense wrapping and attribution
As discussed, here's the version with slightly more inclusive copyright
attribution.  Also I've updated the line wrapping to be a max of 75
characters wide for easy inclusion in source files.  Finally, put the
tiny intro into license.rst so that bsd.txt contains the license text
only, exactly as it will be placed in source files.