Further split surface storage refactoring
After more thought and some experiments, it seems that using a tree for
surface storage was premature.  In fact, a simple flat array of buckets
seems to work with comparable efficiency, as long as care is taken to
reduce the work done due to expired geometry references.  In particular,
it's important to purge expired references before making a heap out of
the references stored in a bucket.

Since the tree stuff has been abandoned for this usage for now, it's
been moved into a new file in anticipation of being useful later.