Refactor file output to do incremental saves.
This allows us to avoid wasting memory buffering the image during
rendering - instead we get rid of it to disk on the fly.  An ImageFile
class was created to hold open image files and handle the actual writing
through the TIFF library.

Note that images with non-factor-of-16 buckets will fail to render in
this revision, since TIFF only accepts tiles of that size natively, so
we'd need an extra layer of buffering to support these.