Region allocator for pipelined memory resources
This is just a cleaned up version of the region allocator which was
prototyped quite a while ago.  It's useful for allocating short-lived
piplined resources efficiently and in a way that can be released back to
the system without causing heap fragmentation.  Because of that, it
provides a way to try to understand the memory usage patterns in depth.