Log memory usage in render & clean up MemoryLog
This change allows heap memory to be logged during rendering if
AQSIS_MEMORY_DEBUGGING is defined in the build.  To speed up the build I
cleaned up memdebug.h by moving some implementation details to
memdebug.cpp.  Also included are some enhancements to the build so that
unit tests can be turned on or off.