Make RegionAllocator use mmap/munmap.
RegionAllocator is designed to allocate memory in quite large blocks, so
it's useful to get these directly from the OS rather than using the
system heap.  If we get memory from the OS using mmap() we can release
the blocks back to the system directly with munmap() when necessary.
This prevents the large blocks from causing memory fragmentation on the
system heap.