Deferred fragment allocation and optional filtering
This change defers allocation of fragments until it's actually needed,
reducing memory usage.  It also turns off the filtering stage where
possible (ie, in the case that no valid fragments exist for a filter
block, in which case we just use the default fragment.)  This
optimization is quite effective in cases where the filtering is
expensive.