Implement separable filtering for buckets.
For large filter regions this is an important optimization because it
reduces the filtering cost from O(N^2) to O(N) where N is the filter
width in samples, at least for separable filters.  If we don't do this,
filtering costs can become a large portion of the render time when the
filter width is large and the geometry is particularly simple.