Basic occlusion culling
Here's a dubiously efficient (but simple for now) version of occlusion
checking: for a given bound, iterate over all the samples which it
touches, and check if any of them are further away than the closest part
of the bound.  Apart from that, here I've added occlusion stats, fixed
some missing stat state changes and improved the way the stats are
presented a bit.