Occlusion tree work.
This is a more or less working version of a quadtree occlusion culling
code, but not integrated into the renderer yet.  The idea is to
construct a pyramid of depth values, where each level of the tree
contains the maximum of all the nodes below it.  Occlusion culling
queries are then extremely fast.  However, at this stage it's not clear
that using this is the best method in all cases because maintaining the
tree data structure is expensive.