Multithreaded bucket rendering with boost.thread
This change does custom bucket scheduling using boost.thread, replacing
the temporary openmp hack.  This seems to get rid of some odd scheduling
pathologies, and results in a nice threaded speedup of 40% or so with
two threads!  Clearly there's some work to do to make this closer to a
linear speedup; it seems likely that the nonlinearity is due to lock
contention in SplitStore::insert().

Details:
* A new command line option -t or --threads has been added to all the
  number of threads to be specified.
* Stats have been disabled in multithreading mode for now, since we need
  per-thread stat support.
* Memory logging is also disabled for now.  It should probably go in the
  bucket scheduler or something...