Overhaul surface storage for threading performance
As it turned out, the surface storage method previously adopted had
significant thread contention issues which were difficult to solve, as
well as some subtle almost-bugs which appear only in the presence of
concurrency.  This change implements an alternative in which we store
the surface splitting tree directly in the geometry holder objects.
Each bucket stores only the initial geometry provided via the API, and
we get to the child geometry split in previous buckets by traversing the
splitting tree.

In general it seems that doing split storage this way really improves
the threading performance (and also the single-threaded performance in
some cases).  For example, the cubescene.rib test (without DoF, which is
currently broken) has a 45% reduction in render time for two threads.

Details:
* Revamp split/dice machinary, including bucket storage, rendering loop
  and tessellation code for major change discussed above.  Stats are
  even more broken by these changes, need to fix them shortly.
* Fix minor bug in Timer usage - timers were started at creation.