Bugfix 2978546: "SHADEOP_TABLE" missing C extern tag
This makes sure that the extern "C" tag is added to prevent name
mangling of the shadeop table.  To avoid warnings with gcc, we declare
the extern "C" on a separate line.

It also cleans up SqShadeOp to contain const char* strings in order to
avoid warnings with recent gcc versions.

Note that this isn't tested on windows yet, though I tried to be
careful.