Move tessellation lock to beginning of tessellation.
This means that we can safely delete the geometry at the end of the
tessellation step because no other thread can be tessellating at the
same time.  There seems to be no real advantage to allowing other
threads to avoid blocking on the tessellation here, since the
alternative is to redo the tessellation work already being done in
another thread.