Set thread affinity when multithreading
This gives a small performance boost when the thread workload is well
distributed.  It also makes timing results much more consistent by
telling the OS scheduler not to shift threads around.  It's possible
that this is a bad move for general workloads but for now it makes
testing a bit easier.