Bugfix #2925557: Pisql locks up if you zoom in too far.
This bug was due to an unfortunate choice by the FLTK designers to use
short integers to store the width, height and position of widgets, so we
can't make the image size any larger than SHRT_MAX or we'll get
overflow.  To fix this in full we'd need to abandon the current design
where the image viewing widget is resized inside a scrollable container.
Since this is a lot of work for only a little added functionality, this
patch simply clamps the image scale for now.