Reduced memory usage for split surfaces
This change adds a reference count to the geometry holder so that split
geometry can be deallocated as soon as possible.  The reference count is
initialized to the number of buckets that the bound crosses.  Whenever a
bucket is finished using a piece of geometry, the refcount is
decremented and the geometry is deleted when it reaches zero.  Without
this scheme, split geometry is never deleted until all the buckets which
the root bound touches are rendered, leading to excessive and unscalable
memory usage.

Details:
* Add extra atomic refcount to GeomHolder to hold the number of
  unrendered buckets touched by the bound.
* Refactor SplitStore::Bucket to remove the queuing of geometry and
  place this in the auxiliary object, GeometryQueue.  The buckets
  themselves are now simple structs.
* Implement geometry lifetime management in GeometryQueue, as described
  above.