Fix GeomHolder bucket refcounting for occlusion
With a split tree-based geometry storage method, a reference count needs
to be included for each piece of geometry, recording how many unrendered
buckets the geometry is still needed for.  When the refcount decreases
to zero, the associated memory can be reclaimed.  Occlusion culling
makes this more difficult, because occlusion of a piece of geometry may
happen based on a parent further up the split tree.  This means we need
to record each bucket in which a piece of geometry is occluded until it
is either fully occlusion culled from all buckets, or split.  If it is
split, we pass the record of which buckets the parent has already been
occluded in down to the children.  The children use the occlusion record
to calculate the correct initial reference count.

Details:
* Create a new member bitvector, GeomHolder::m_occludedBuckets, which
  holds a record of which buckets the geometry has been occluded in.
  This is allocated on demand to avoid wasting memory unnecessarily.
* Various stuff in GeomHolder, Renderer, and TessellationContextImpl to
  cater for maintaining the occlusion record and its interaction with
  the geometry refcount.

Note that the grids should also have a bucket refcount because their
bounds can be significantly smaller than the associated parent surface
due to displacement bound expansion.  This isn't implemented yet.