Enable bucket refcounting for grids
This adds a count of the number of remaining buckets touched by each
grid.  This allows us to delete grids immediately after they have been
rasterized, which is important in keeping memory usage to a minimum,
particularly for large displacement bounds.  (For example, with the
cubescene.rib test renderered at 2000x2000 without DoF, the memory usage
decreases by 40% with this optimization.)