Add a small number of Intel TBB atomic headers
It's really good to have a full set of atomic operations for some
algorithms.  In the absence of C++0x std::atomic<>, I've chosen to
include the atomic.h and associated support headers from Intel's thread
building blocks library.  (Actually, these come by way of the OIIO
project's stripped down version of TBB.)

The Intel TBB license is GPL2 with the libstdc++ exception clause which
makes it BSD-compatible.  Ideally we would just use tbb directly, but
it's a very heavy dependency for the small number of features I'm using
here.