Fltk interactive demo, first steps
This is an interactive application written in fltk to demonstrate the
basic features (and speed!) of the new core.  So far there's not much
"interactive" about it, apart from a fltk window which shows the render
result.

This change also includes a crude hack to pick up Display objects which
are passed in via the Ri::Renderer interface.  Such displays are
detected by setting the type field in the Display() call to the string
"__Display_instance__", and including a "int instance" parameter.  Using
int here is clearly wrong, but more work is required to implement a
basic "pointer" type in the interface.