Improve eyesplit handling to avoid hangs and crashes
The presence of eye splits caused the renderer to hang or crash in
certian circumstances, for several reasons:

1) The displacement bound was added after checking for the max eye
splits condition rather than before.  This meant that objects with large
displacement bound would get split indefinitely rather than being
discarded eventually.

2) The projection was used to determine the split direction even for
geometry which spans the epsilon plane, but this isn't reliable.  In
this case the right thing to do is to force the geometry to split
(rather than dice), and get the split direction to alternate in the hope
that it will result in pieces which can eventually be either culled or
rendered.  This is implemented via the new forceSplit parameter to
Geometry::tessellate()

3) The corners of the projected bound are unreliable for bounding
geometry which spans the epsilon plane - instead the resulting bound
is either infinite or semi-infinite.  We work around this by setting the
bound to the image extent for any such pieces of geometry.

As an added precaution, the near clipping plane for the interactive demo
was also set to something more sane than the default of FLT_EPSILON.
This is actually a requirement for any geometry which has a nonzero
displacement bound.